tests/wpt/web-platform-tests/web-animations/interfaces/AnimationEffectTiming/delay.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>delay tests</title>
<link rel="help" href="https://w3c.github.io/web-animations/#dom-animationeffecttiming-delay">
<link rel="author" title="Daisuke Akatsuka" href="mailto:daisuke@mozilla-japan.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../testcommon.js"></script>
<body>
<div id="log"></div>
<script>
'use strict';

test(function(t) {
  var anim = createDiv(t).animate(null);
  assert_equals(anim.effect.timing.delay, 0);
}, 'Test default value');

test(function(t) {
  var div = createDiv(t);
  var anim = div.animate({ opacity: [ 0, 1 ] }, 100);
  anim.effect.timing.delay = 100;
  assert_equals(anim.effect.timing.delay, 100, 'set delay 100');
  assert_equals(anim.effect.getComputedTiming().delay, 100,
                  'getComputedTiming() after set delay 100');
}, 'set delay 100');

test(function(t) {
  var div = createDiv(t);
  var anim = div.animate({ opacity: [ 0, 1 ] }, 100);
  anim.effect.timing.delay = -100;
  assert_equals(anim.effect.timing.delay, -100, 'set delay -100');
  assert_equals(anim.effect.getComputedTiming().delay, -100,
                'getComputedTiming() after set delay -100');
}, 'set delay -100');

test(function(t) {
  var div = createDiv(t);
  var anim = div.animate({ opacity: [ 0, 1 ] }, 100);
  anim.effect.timing.delay = 100;
  assert_equals(anim.effect.getComputedTiming().progress, null);
  assert_equals(anim.effect.getComputedTiming().currentIteration, null);
}, 'Test adding a positive delay to an animation without a backwards fill ' +
   'makes it no longer active');

test(function(t) {
  var div = createDiv(t);
  var anim = div.animate({ opacity: [ 0, 1 ] },
                         { fill: 'both',
                           duration: 100 });
  anim.effect.timing.delay = -50;
  assert_equals(anim.effect.getComputedTiming().progress, 0.5);
}, 'Test seeking an animation by setting a negative delay');

test(function(t) {
  var div = createDiv(t);
  var anim = div.animate({ opacity: [ 0, 1 ] },
                         { fill: 'both',
                           duration: 100 });
  anim.effect.timing.delay = -100;
  assert_equals(anim.effect.getComputedTiming().progress, 1);
  assert_equals(anim.effect.getComputedTiming().currentIteration, 0);
}, 'Test finishing an animation using a large negative delay');

test(function(t) {
  var div = createDiv(t);
  var anim = div.animate(null);
  for (let invalid of [NaN, Infinity]) {
    assert_throws({ name: 'TypeError' }, function() {
      anim.effect.timing.delay = invalid;
    }, 'setting ' + invalid);
    assert_throws({ name: 'TypeError' }, function() {
      div.animate({}, { delay: invalid });
    }, 'animate() with ' + invalid);
  }
}, 'Setting invalid values should throw TypeError');

</script>
</body>
